+++ Services/CDN/CDN.Application/Features/CategoryFeature/Commands/UpdateCategory/UpdateCategoryCommand.cs
@@ -1,11 +1,10 @@
 ﻿using MediatR;
-using System;
 
 namespace CDN.Application.Features.CategoryFeature.Commands.UpdateCategory
 {
     public class UpdateCategoryCommand : IRequest
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; }
         public string Path { get; set; }
         public bool IsActive { get; set; }
