+++ Services/CDN/CDN.Application/Features/FileFeature/Commands/UploadFile/UploadFileCommand.cs
@@ -5,12 +5,12 @@ namespace CDN.Application.Features.FileFeature.Commands.UploadFile
 {
     public class UploadFileCommand : IRequest<string>
     {
-        public long CategoryId { get; set; }
+        public int CategoryId { get; set; }
         public IFormFile FormFile { get; set; }
         public string TagName { get; set; }
         public string FileName { get; set; }
 
-        public UploadFileCommand(long categoryId, IFormFile formFile, string tagName, string fileName)
+        public UploadFileCommand(int categoryId, IFormFile formFile, string tagName, string fileName)
         {
             CategoryId = categoryId;
             FormFile = formFile;
